+++ BackEnd/src/main/java/com/example/server/dao/RoleDao.java
@@ -7,10 +7,10 @@ import com.example.server.entity.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-public interface RoleDao extends JpaRepository<Role, Integer> {
+public interface RoleDao extends JpaRepository<Role, Long> {
     Role findRoleByName(String name);
     
-    Role findRoleById(int id);
+    Role findRoleById(Long id);
 
     @Query("select r from Role r " +
     "where r.is_deleted = 0 " +
